[Extensions] 생성, 해당 파일 주 객체와 연관이 크게 없는 객체의 확장 코드를 Extensions파일로 이동
