+++ StoreApp/StoreApp/Models/StoreItems.swift
@@ -54,11 +54,6 @@ class StoreItems {
     }
 }
 
-extension NSNotification.Name {
-    static let storeItemsWillUpdate = NSNotification.Name("storeItemsDidUpdate")
-    static let rowWillReload = NSNotification.Name("rowWillReload")
-}
-
 extension JSONDataFetcher {
     func fetchStoreItems(fileName: String, completion: @escaping ([StoreItem]) -> Void) {
         let commonURL = "https://h3rb9c0ugl.execute-api.ap-northeast-2.amazonaws.com/develop/baminchan/"
